refactor: Rename Exceptions class
