+++ src/main/java/com/spotify/confidence/Confidence.java
@@ -9,11 +9,11 @@ import com.google.common.annotations.VisibleForTesting;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Maps;
 import com.google.common.io.Closer;
-import com.spotify.confidence.ConfidenceExceptions.IllegalValuePath;
-import com.spotify.confidence.ConfidenceExceptions.IllegalValueType;
-import com.spotify.confidence.ConfidenceExceptions.IncompatibleValueType;
-import com.spotify.confidence.ConfidenceExceptions.ValueNotFound;
 import com.spotify.confidence.ConfidenceUtils.FlagPath;
+import com.spotify.confidence.Exceptions.IllegalValuePath;
+import com.spotify.confidence.Exceptions.IllegalValueType;
+import com.spotify.confidence.Exceptions.IncompatibleValueType;
+import com.spotify.confidence.Exceptions.ValueNotFound;
 import com.spotify.confidence.shaded.flags.resolver.v1.ResolveFlagsResponse;
 import com.spotify.confidence.shaded.flags.resolver.v1.ResolvedFlag;
 import io.grpc.ManagedChannel;
