+++ src/main/java/com/spotify/confidence/ConfidenceFeatureProvider.java
@@ -4,8 +4,8 @@ import static com.spotify.confidence.ConfidenceUtils.FlagPath.getPath;
 import static com.spotify.confidence.OpenFeatureUtils.getValueForPath;
 
 import com.google.protobuf.Struct;
-import com.spotify.confidence.ConfidenceExceptions.IllegalValuePath;
 import com.spotify.confidence.ConfidenceUtils.FlagPath;
+import com.spotify.confidence.Exceptions.IllegalValuePath;
 import com.spotify.confidence.shaded.flags.resolver.v1.ResolveFlagsResponse;
 import com.spotify.confidence.shaded.flags.resolver.v1.ResolvedFlag;
 import dev.openfeature.sdk.EvaluationContext;
